Return to sum-of-errors based dispersion calculation.
