+++ src/mondefense.cpp
@@ -1,4 +1,7 @@
 #include "mondefense.h"
+
+#include "ballistics.h"
+#include "dispersion.h"
 #include "monster.h"
 #include "creature.h"
 #include "damage.h"
@@ -102,7 +105,7 @@ void mdefense::acidsplash( monster &m, Creature *const source,
     prj.impact.add_damage( DT_ACID, rng( 1, 3 ) );
     for( size_t i = 0; i < num_drops; i++ ) {
         const tripoint &target = random_entry( pts );
-        m.projectile_attack( prj, target, 1200 );
+        projectile_attack( prj, m.pos(), target, { 1200 } );
     }
 
     if( g->u.sees( m.pos() ) ) {
