+++ src/ranged.cpp
@@ -1,6 +1,8 @@
 #include "ranged.h"
 
+#include "ballistics.h"
 #include "cata_utility.h"
+#include "dispersion.h"
 #include "game.h"
 #include "map.h"
 #include "debug.h"
@@ -8,12 +10,12 @@
 #include "line.h"
 #include "skill.h"
 #include "rng.h"
-#include "projectile.h"
 #include "item.h"
 #include "options.h"
 #include "action.h"
 #include "input.h"
 #include "messages.h"
+#include "projectile.h"
 #include "sounds.h"
 #include "translations.h"
 #include "monster.h"
@@ -29,8 +31,6 @@
 #include <string>
 #include <cmath>
 
-const float sqrt2 = std::sqrt( 2.0f );
-
 const skill_id skill_throw( "throw" );
 const skill_id skill_gun( "gun" );
 const skill_id skill_driving( "driving" );
@@ -38,7 +38,6 @@ const skill_id skill_dodge( "dodge" );
 const skill_id skill_launcher( "launcher" );
 
 const efftype_id effect_on_roof( "on_roof" );
-const efftype_id effect_bounced( "bounced" );
 const efftype_id effect_hit_by_player( "hit_by_player" );
 
 static projectile make_gun_projectile( const item &gun );
@@ -46,7 +45,6 @@ int time_to_fire( const Character &p, const itype &firing );
 static void cycle_action( item& weap, const tripoint &pos );
 void make_gun_sound_effect(player &p, bool burst, item *weapon);
 extern bool is_valid_in_w_terrain(int, int);
-void drop_or_embed_projectile( const dealt_projectile_attack &attack );
 
 struct aim_type {
     std::string name;
@@ -55,69 +53,6 @@ struct aim_type {
     bool has_threshold;
     int threshold;
 };
-
-dealt_projectile_attack Creature::projectile_attack( const projectile &proj, const tripoint &target,
-                                                     double dispersion )
-{
-    return projectile_attack( proj, pos(), target, dispersion );
-}
-
-size_t blood_trail_len( int damage )
-{
-    if( damage > 50 ) {
-        return 3;
-    } else if( damage > 20 ) {
-        return 2;
-    } else if ( damage > 0 ) {
-        return 1;
-    }
-    return 0;
-}
-
-// projectile_attack handles a single shot aimed at target_arg. It handles all the randomization of
-// where the shot goes versus where it was aimed.  The dispersion angle passed to projectile_attack
-// represents the possible error due to imperfections in the weapon and the steadiness of the
-// shooter's aim; for the same weapon / shooter / recoil you'll have the same dispersion.
-
-// The path each shot takes has some random error.  This is modelled as a normal distribution
-// centered around the aiming point, or - equivalently - as an error angle (shot_dispersion) that is
-// centered around zero. A shot_dispersion of zero is a perfectly-on-target shot. shot_dispersion,
-// unlike dispersion, is a signed value, the error might be to the left (negative) or right
-// (positive).
-
-// The standard deviation of the normal distribution is (dispersion / dispersion_sigmas). With the
-// current value of dispersion_sigmas (2.4) this means that the dispersion value is 2.4 standard
-// deviations and 98% of all shots will lie within the dispersion angle.
-
-// Both dispersion and shot_dispersion are measured in arcminutes (1/60th of a degree).
-
-// Increasing dispersion_sigmas will decrease the standard deviation, shots will be "tighter" around
-// the actual target, the overall chance to hit increases.
-
-// Allowing a hit on the target just because you hit the right tile doesn't work balance-wise as
-// it's too big of a target if you also want to keep miss angles within a reasonable range. So there
-// is a second parameter, occupied_tile_fraction, which says "how wide" the target is. You have to
-// get at least this close with your shot, in tiles (missed_by_tiles) to have a chance to
-// hit. Increasing occupied_tile_fraction increases the chance to hit.
-
-// You should not set occupied_tile_fraction > 1.0 as that would mean that hitting an adjacent tile
-// is enough to hit the critter, and the code that follows the shot trajectory looking for hits is
-// not capable of handling that.
-
-// Once you are close enough with the shot for a hit, exactly how close influences the quality of
-// the hit.  missed_by represents how good the hit is on a scale of 0 to 1; 0.0 is a perfect hit;
-// 0.99999.. is the worst possible shot that still connects (where you got just barely within
-// occupied_tile_fraction tiles of where you aimed and just clipped the edge of your target)
-
-// dispersion_sigmas and occupied_tile_fraction can be tuned together to affect the spread of misses
-// without affecting hit or hit quality chances, e.g. if you increase dispersion_sigmas by 10% and
-// also decrease occupied_tile_fraction by 10%, then the cone of misses will get 10% tighter but the
-// overall hit and hit quality chances are unchanged.
-
-// these parameters balance the hit/miss chance:
-
-/** How many standard deviations does the dispersion passed to Creature::projectile_attack represent? */
-const double dispersion_sigmas = 2.4;
 
 static double occupied_tile_fraction( m_size target_size )
 {
@@ -137,364 +72,9 @@ static double occupied_tile_fraction( m_size target_size )
     return 0.5;
 }
 
-projectile_attack_aim Creature::projectile_attack_roll( double dispersion, double range, double target_size ) const
-{
-    projectile_attack_aim aim;
-
-    // dispersion is a measure of the dispersion of shots due to the gun + shooter characteristics
-    // i.e. it is independent of any particular shot
-
-    // shot_dispersion is the actual dispersion for this particular shot, i.e.
-    // the error angle between where the shot was aimed and where this one actually went
-    // NB: some cases pass dispersion == 0 for a "never misses" shot e.g. bio_magnet,
-    aim.dispersion = dispersion > 0 ? normal_roll( 0.0, dispersion / dispersion_sigmas ) : 0;
-
-    // an isosceles triangle is formed by the intended and actual target tiles
-    aim.missed_by_tiles = iso_tangent( range, aim.dispersion );
-
-    // fraction we missed a monster target by (0.0 = perfect hit, 1.0 = miss)
-    if( target_size > 0.0 ) {
-        aim.missed_by = std::min( 1.0, aim.missed_by_tiles / target_size );
-    } else {
-        // Special case 0 size targets, just to be safe from 0.0/0.0 NaNs
-        aim.missed_by = 1.0;
-    }
-
-    return aim;
-}
-
 double Creature::ranged_target_size() const
 {
     return occupied_tile_fraction( get_size() );
-}
-
-dealt_projectile_attack Creature::projectile_attack( const projectile &proj_arg, const tripoint &source,
-                                                     const tripoint &target_arg, double dispersion )
-{
-    const bool do_animation = get_option<bool>( "ANIMATIONS" );
-
-    double range = rl_dist( source, target_arg );
-
-    Creature *target_critter = g->critter_at( target_arg );
-    double target_size = target_critter != nullptr ?
-                             target_critter->ranged_target_size() :
-                             g->m.ranged_target_size( target_arg );
-    projectile_attack_aim aim = projectile_attack_roll( dispersion, range, target_size );
-
-    // TODO: move to-hit roll back in here
-
-    dealt_projectile_attack attack {
-        proj_arg, nullptr, dealt_damage_instance(), source, aim.missed_by
-    };
-
-    if( source == target_arg ) { // No suicidal shots
-        debugmsg( "%s tried to commit suicide.", get_name().c_str() );
-        return attack;
-    }
-
-    projectile &proj = attack.proj;
-    const auto &proj_effects = proj.proj_effects;
-
-    const bool stream = proj_effects.count("STREAM") > 0 ||
-                        proj_effects.count("STREAM_BIG") > 0 ||
-                        proj_effects.count("JET") > 0;
-    const bool no_item_damage = proj_effects.count( "NO_ITEM_DAMAGE" ) > 0;
-    const bool do_draw_line = proj_effects.count( "DRAW_AS_LINE" ) > 0;
-    const bool null_source = proj_effects.count( "NULL_SOURCE" ) > 0;
-    // Determines whether it can penetrate obstacles
-    const bool is_bullet = proj_arg.speed >= 200 && std::any_of( proj_arg.impact.damage_units.begin(),
-                           proj_arg.impact.damage_units.end(),
-                           []( const damage_unit &dam )
-    {
-        return dam.type == DT_CUT;
-    } );
-
-    // If we were targetting a tile rather than a monster, don't overshoot
-    // Unless the target was a wall, then we are aiming high enough to overshoot
-    const bool no_overshoot = proj_effects.count( "NO_OVERSHOOT" ) ||
-                              ( g->critter_at( target_arg ) == nullptr && g->m.passable( target_arg ) );
-
-    double extend_to_range = no_overshoot ? range : proj_arg.range;
-
-    tripoint target = target_arg;
-    std::vector<tripoint> trajectory;
-    if( aim.missed_by_tiles >= 1.0 ) {
-        // We missed enough to target a different tile
-        double dx = target_arg.x - source.x;
-        double dy = target_arg.y - source.y;
-        double rad = atan2( dy, dx );
-
-        // cap wild misses at +/- 30 degrees
-        rad += std::max( std::min( ARCMIN( aim.dispersion ), DEGREES( 30 ) ), DEGREES( -30 ) );
-
-        // @todo This should also represent the miss on z axis
-        const int offset = std::min<int>( range, sqrtf( aim.missed_by_tiles ) );
-        int new_range = no_overshoot ?
-                            range + rng( -offset, offset ) :
-                            rng( range - offset, proj_arg.range );
-        new_range = std::max( new_range, 1 );
-
-        target.x = source.x + roll_remainder( new_range * cos( rad ) );
-        target.y = source.y + roll_remainder( new_range * sin( rad ) );
-
-        if( target == source ) {
-            target.x = source.x + sgn( dx );
-            target.y = source.y + sgn( dy );
-        }
-
-        // Don't extend range further, miss here can mean hitting the ground near the target
-        range = rl_dist( source, target );
-        extend_to_range = range;
-
-        sfx::play_variant_sound( "bullet_hit", "hit_wall", sfx::get_heard_volume( target ), sfx::get_heard_angle( target ));
-        // TODO: Z dispersion
-        // If we missed, just draw a straight line.
-        trajectory = line_to( source, target );
-    } else {
-        // Go around obstacles a little if we're on target.
-        trajectory = g->m.find_clear_path( source, target );
-    }
-
-    add_msg( m_debug, "%s proj_atk: dispersion: %.2f; aim.dispersion: %.2f", disp_name().c_str(), dispersion, aim.dispersion );
-
-    add_msg( m_debug, "missed_by_tiles: %.2f; missed_by: %.2f; target (orig/hit): %d,%d,%d/%d,%d,%d", aim.missed_by_tiles, aim.missed_by,
-             target_arg.x, target_arg.y, target_arg.z,
-             target.x, target.y, target.z );
-
-    // Trace the trajectory, doing damage in order
-    tripoint &tp = attack.end_point;
-    tripoint prev_point = source;
-
-    trajectory.insert( trajectory.begin(), source ); // Add the first point to the trajectory
-
-    static emit_id muzzle_smoke( "emit_smoke_plume" );
-    if( proj_effects.count( "MUZZLE_SMOKE" ) ) {
-        g->m.emit_field( trajectory.front(), muzzle_smoke );
-    }
-
-    if( !no_overshoot && range < extend_to_range ) {
-        // Continue line is very "stiff" when the original range is short
-        // @todo Make it use a more distant point for more realistic extended lines
-        std::vector<tripoint> trajectory_extension = continue_line( trajectory,
-                                                                    extend_to_range - range );
-        trajectory.reserve( trajectory.size() + trajectory_extension.size() );
-        trajectory.insert( trajectory.end(), trajectory_extension.begin(), trajectory_extension.end() );
-    }
-    // Range can be 0
-    size_t traj_len = trajectory.size();
-    while( traj_len > 0 && rl_dist( source, trajectory[traj_len-1] ) > proj_arg.range ) {
-        --traj_len;
-    }
-
-    // If this is a vehicle mounted turret, which vehicle is it mounted on?
-    const vehicle *in_veh = has_effect( effect_on_roof ) ?
-                            g->m.veh_at( pos() ) : nullptr;
-
-    const float projectile_skip_multiplier = 0.1;
-    // Randomize the skip so that bursts look nicer
-    int projectile_skip_calculation = range * projectile_skip_multiplier;
-    int projectile_skip_current_frame = rng( 0, projectile_skip_calculation );
-    bool has_momentum = true;
-
-    for( size_t i = 1; i < traj_len && ( has_momentum || stream ); ++i ) {
-        prev_point = tp;
-        tp = trajectory[i];
-
-        if( ( tp.z > prev_point.z && g->m.has_floor( tp ) ) ||
-            ( tp.z < prev_point.z && g->m.has_floor( prev_point ) ) ) {
-            // Currently strictly no shooting through floor
-            // TODO: Bash the floor
-            tp = prev_point;
-            traj_len = --i;
-            break;
-        }
-        // Drawing the bullet uses player g->u, and not player p, because it's drawn
-        // relative to YOUR position, which may not be the gunman's position.
-        if( do_animation && !do_draw_line ) {
-            // TODO: Make this draw thrown item/launched grenade/arrow
-            if( projectile_skip_current_frame >= projectile_skip_calculation ) {
-                g->draw_bullet(g->u, tp, (int)i, trajectory, stream ? '#' : '*');
-                projectile_skip_current_frame = 0;
-                // If we missed recalculate the skip factor so they spread out.
-                projectile_skip_calculation = std::max( (size_t)range, i ) * projectile_skip_multiplier;
-            } else {
-                projectile_skip_current_frame++;
-            }
-        }
-
-        if( in_veh != nullptr ) {
-            int part;
-            vehicle *other = g->m.veh_at( tp, part );
-            if( in_veh == other && other->is_inside( part ) ) {
-                continue; // Turret is on the roof and can't hit anything inside
-            }
-        }
-
-        Creature *critter = g->critter_at( tp );
-        if( critter == this ) {
-            // No hitting self with "weird" attacks
-            critter = nullptr;
-        }
-
-        monster *mon = dynamic_cast<monster *>(critter);
-        // ignore non-point-blank digging targets (since they are underground)
-        if( mon != nullptr && mon->digging() &&
-            rl_dist( pos(), tp ) > 1) {
-            critter = nullptr;
-            mon = nullptr;
-        }
-
-        // Reset hit critter from the last iteration
-        attack.hit_critter = nullptr;
-
-        // If we shot us a monster...
-        // TODO: add size effects to accuracy
-        // If there's a monster in the path of our bullet, and either our aim was true,
-        //  OR it's not the monster we were aiming at and we were lucky enough to hit it
-        double cur_missed_by = aim.missed_by;
-
-        // unintentional hit on something other than our actual target
-        // don't re-roll for the actual target, we already decided on a missed_by value for that
-        // at the start, misses should stay as misses
-        if( critter != nullptr && tp != target_arg ) {
-            // Unintentional hit
-            cur_missed_by = std::max( rng_float( 0.1, 1.5 - aim.missed_by ) / critter->ranged_target_size(), 0.4 );
-        }
-
-        if( critter != nullptr && cur_missed_by < 1.0 ) {
-            if( in_veh != nullptr && g->m.veh_at( tp ) == in_veh && critter->is_player() ) {
-                // Turret either was aimed by the player (who is now ducking) and shoots from above
-                // Or was just IFFing, giving lots of warnings and time to get out of the line of fire
-                continue;
-            }
-            attack.missed_by = cur_missed_by;
-            critter->deal_projectile_attack( null_source ? nullptr : this, attack );
-            // Critter can still dodge the projectile
-            // In this case hit_critter won't be set
-            if( attack.hit_critter != nullptr ) {
-                const size_t bt_len = blood_trail_len( attack.dealt_dam.total_damage() );
-                if( bt_len > 0 ) {
-                    const tripoint &dest = move_along_line( tp, trajectory, bt_len );
-                    g->m.add_splatter_trail( critter->bloodType(), tp, dest );
-                }
-                sfx::do_projectile_hit( *attack.hit_critter );
-                has_momentum = false;
-            } else {
-                attack.missed_by = aim.missed_by;
-            }
-        } else if( in_veh != nullptr && g->m.veh_at( tp ) == in_veh ) {
-            // Don't do anything, especially don't call map::shoot as this would damage the vehicle
-        } else {
-            g->m.shoot( tp, proj, !no_item_damage && tp == target );
-            has_momentum = proj.impact.total_damage() > 0;
-        }
-
-        if( ( !has_momentum || !is_bullet ) && g->m.impassable( tp ) ) {
-            // Don't let flamethrowers go through walls
-            // TODO: Let them go through bars
-            traj_len = i;
-            break;
-        }
-    } // Done with the trajectory!
-
-    if( do_animation && do_draw_line && traj_len > 2 ) {
-        trajectory.erase( trajectory.begin() );
-        trajectory.resize( traj_len-- );
-        g->draw_line( tp, trajectory );
-        g->draw_bullet( g->u, tp, int( traj_len-- ), trajectory, stream ? '#' : '*' );
-    }
-
-    if( g->m.impassable(tp) ) {
-        tp = prev_point;
-    }
-
-    drop_or_embed_projectile( attack );
-
-    apply_ammo_effects( tp, proj.proj_effects );
-    const auto &expl = proj.get_custom_explosion();
-    if( expl.power > 0.0f ) {
-        g->explosion( tp, proj.get_custom_explosion() );
-    }
-
-    // TODO: Move this outside now that we have hit point in return values?
-    if( proj.proj_effects.count( "BOUNCE" ) ) {
-        for( size_t i = 0; i < g->num_zombies(); i++ ) {
-            monster &z = g->zombie(i);
-            if( z.is_dead() ) {
-                continue;
-            }
-            // search for monsters in radius 4 around impact site
-            if( rl_dist( z.pos(), tp ) <= 4 &&
-                g->m.sees( z.pos(), tp, -1 ) ) {
-                // don't hit targets that have already been hit
-                if( !z.has_effect( effect_bounced ) ) {
-                    add_msg(_("The attack bounced to %s!"), z.name().c_str());
-                    z.add_effect( effect_bounced, 1 );
-                    projectile_attack( proj, tp, z.pos(), dispersion );
-                    sfx::play_variant_sound( "fire_gun", "bio_lightning_tail", sfx::get_heard_volume(z.pos()), sfx::get_heard_angle(z.pos()));
-                    break;
-                }
-            }
-        }
-    }
-
-    return attack;
-}
-
-double player::gun_current_range( const item& gun, double penalty, unsigned chance, double accuracy ) const
-{
-    if( !gun.is_gun() || !gun.ammo_sufficient() ) {
-        return 0;
-    }
-
-    if( chance == 0 ) {
-        return gun.gun_range( this );
-    }
-
-    // calculate base dispersion
-    double dispersion = get_weapon_dispersion( gun, RANGE_SOFT_CAP ) + ( penalty < 0 ? recoil_total() : penalty );
-    double range = ranged::effective_range( dispersion, chance, accuracy, occupied_tile_fraction( MS_MEDIUM ) );
-    range = std::min( range, ( double )gun.gun_range( this ) );
-    return range;
-}
-
-double player::thrown_current_range( const item& to_throw, unsigned chance,
-                                     double accuracy, Creature *target ) const
-{
-    double dispersion = throwing_dispersion( to_throw, target );
-    float target_size = occupied_tile_fraction( target != nullptr ? target->get_size() : MS_MEDIUM );
-    double range = ranged::effective_range( dispersion, chance, accuracy, target_size );
-    range = std::min( range, ( double )throw_range( to_throw ) );
-    return range;
-}
-
-namespace ranged {
-
-double effective_range( double dispersion, unsigned chance, double accuracy, double target_size )
-{
-    double sigma = dispersion / dispersion_sigmas;
-
-    // angle such that chance% of dispersion rolls are <= that angle
-    double max_dispersion = erfinv( chance / 100.0 ) * sqrt2 * sigma;
-
-    // required iso_tangent value
-    // Assume we're displaying ranges against humans and human-sized zombies
-    double missed_by_tiles = accuracy * target_size;
-
-    // work backwards to range
-    //   T = (2*D**2 * (1 - cos V)) ** 0.5   (from iso_tangent)
-    //   D = (0.5*T**2 / (1 - cos V)) ** 0.5
-    double range = sqrt( 0.5 * missed_by_tiles * missed_by_tiles / ( 1 - cos( ARCMIN( max_dispersion ) ) ) );
-
-    if( range > RANGE_SOFT_CAP ) {
-        // This is a special case that is a pain to calculate since dispersion depends on range
-        // Just interpolate linearly - the actual function is linear enough
-        return lerp_clamped( RANGE_SOFT_CAP, RANGE_HARD_CAP, 1.0 - dispersion / 180.0 );
-    }
-
-    return range;
-}
-
 }
 
 int player::gun_engagement_moves( const item &gun ) const
@@ -627,10 +207,14 @@ int player::fire_gun( const tripoint &target, int shots, item& gun )
             break;
         }
 
-        double dispersion = get_weapon_dispersion( gun, rl_dist( pos(), aim ) ) + recoil_total();
+        dispersion_sources dispersion = get_weapon_dispersion( gun, rl_dist( pos(), aim ) );
+        dispersion.add_range( recoil_total() );
         int range = rl_dist( pos(), aim );
 
-        auto shot = projectile_attack( make_gun_projectile( gun ), aim, dispersion );
+        // If this is a vehicle mounted turret, which vehicle is it mounted on?
+        const vehicle *in_veh = has_effect( effect_on_roof ) ? g->m.veh_at( pos() ) : nullptr;
+
+        auto shot = projectile_attack( make_gun_projectile( gun ), pos(), aim, dispersion, this, in_veh );
         curshot++;
 
         int qty = gun.gun_recoil( *this, bipod );
@@ -897,8 +481,8 @@ dealt_projectile_attack player::throw_item( const tripoint &target, const item &
     const float final_xp_mult = range_factor * damage_factor;
 
     Creature *critter = g->critter_at( target, true );
-    const double dispersion = throwing_dispersion( thrown, critter );
-    auto dealt_attack = projectile_attack( proj, target, dispersion );
+    const dispersion_sources dispersion = throwing_dispersion( thrown, critter );
+    auto dealt_attack = projectile_attack( proj, pos(), target, dispersion, this );
 
     const double missed_by = dealt_attack.missed_by;
     if( missed_by <= 0.1 && dealt_attack.hit_critter != nullptr ) {
@@ -1062,7 +646,7 @@ struct confidence_rating {
 
 static int print_ranged_chance( WINDOW *w, int line_number,
                                 const std::vector<confidence_rating> &confidence_config,
-                                double dispersion, double range, double target_size,
+                                dispersion_sources dispersion, double range, double target_size,
                                 double steadiness = 10000.0 )
 {
     const int window_width = getmaxx( w ) - 2; // Window width minus borders.
@@ -1071,7 +655,7 @@ static int print_ranged_chance( WINDOW *w, int line_number,
     // is not doing gaussian integration in their head while aiming.  The result gives the player
     // correct relative measures of chance to hit, and corresponds with the actual distribution at
     // min, max, and mean.
-    const double max_lateral_offset = iso_tangent( range, dispersion );
+    const double max_lateral_offset = iso_tangent( range, dispersion.max() );
     const double confidence = 1 / ( max_lateral_offset / target_size );
 
     bool print_steadiness = steadiness < 1000.0;
@@ -1125,7 +709,9 @@ static int print_aim( const player &p, WINDOW *w, int line_number, item *weapon,
     // Dodge doesn't affect gun attacks
 
     const double range = rl_dist( p.pos(), target.pos() );
-    const double dispersion = p.get_weapon_dispersion( *weapon, range ) + predicted_recoil + p.recoil_vehicle();
+    dispersion_sources dispersion = p.get_weapon_dispersion( *weapon, range );
+    dispersion.add_range( predicted_recoil );
+    dispersion.add_range( p.recoil_vehicle() );
 
     // This is a relative measure of how steady the player's aim is,
     // 0 it is the best the player can do.
@@ -1172,7 +758,7 @@ static int draw_throw_aim( const player &p, WINDOW *w, int line_number,
         target = nullptr;
     }
 
-    const double dispersion = p.throwing_dispersion( *weapon, target );
+    const dispersion_sources dispersion = p.throwing_dispersion( *weapon, target );
     const double range = rl_dist( p.pos(), target_pos );
 
     const double target_size = target != nullptr ? target->ranged_target_size() : 1.0f;
@@ -1883,12 +1469,12 @@ static bool is_driving( const player &p )
 
 
 // utility functions for projectile_attack
-double player::get_weapon_dispersion( const item &obj, float range ) const
+dispersion_sources player::get_weapon_dispersion( const item &obj, float range ) const
 {
-    double dispersion = obj.gun_dispersion();
+    dispersion_sources dispersion = obj.gun_dispersion();
 
     /** @EFFECT_GUN improves usage of accurate weapons and sights */
-    dispersion += 10 * ( MAX_SKILL - std::min( int( get_skill_level( skill_gun ) ), MAX_SKILL ) );
+    dispersion.add_range( 10 * ( MAX_SKILL - std::min( int( get_skill_level( skill_gun ) ), MAX_SKILL ) ) );
 
     if( is_driving( *this ) ) {
         // get volume of gun (or for auxiliary gunmods the parent gun)
@@ -1896,108 +1482,26 @@ double player::get_weapon_dispersion( const item &obj, float range ) const
         const int vol = ( parent ? parent->volume() : obj.volume() ) / 250_ml;
 
         /** @EFFECT_DRIVING reduces the inaccuracy penalty when using guns whilst driving */
-        dispersion += std::max( vol - get_skill_level( skill_driving ), 1 ) * 20;
+        dispersion.add_range( std::max( vol - get_skill_level( skill_driving ), 1 ) * 20 );
     }
 
-    if (has_bionic("bio_targeting")) {
-        dispersion *= 0.75;
+    if( has_bionic( "bio_targeting" ) ) {
+        dispersion.add_multiplier( 0.75 );
     }
 
-    if ((is_underwater() && !obj.has_flag("UNDERWATER_GUN")) ||
+    if( ( is_underwater() && !obj.has_flag( "UNDERWATER_GUN" ) ) ||
         // Range is effectively four times longer when shooting unflagged guns underwater.
-        (!is_underwater() && obj.has_flag("UNDERWATER_GUN"))) {
+        ( !is_underwater() && obj.has_flag( "UNDERWATER_GUN" ) ) ) {
         // Range is effectively four times longer when shooting flagged guns out of water.
-        dispersion *= 4;
+        dispersion.add_multiplier( 4 );
     }
 
     // @todo Scale the range penalty with something (but don't allow it to get below 2.0*range)
     if( range > RANGE_SOFT_CAP ) {
-        dispersion += ( range - RANGE_SOFT_CAP ) * 3.0f;
+        dispersion.add_range( ( range - RANGE_SOFT_CAP ) * 3.0f );
     }
 
-    return std::max( dispersion, 0.0 );
-}
-
-void drop_or_embed_projectile( const dealt_projectile_attack &attack )
-{
-    const auto &proj = attack.proj;
-    const auto &drop_item = proj.get_drop();
-    const auto &effects = proj.proj_effects;
-    if( drop_item.is_null() ) {
-        return;
-    }
-
-    const tripoint &pt = attack.end_point;
-
-    if( effects.count( "SHATTER_SELF" ) ) {
-        // Drop the contents, not the thrown item
-        if( g->u.sees( pt ) ) {
-            add_msg( _("The %s shatters!"), drop_item.tname().c_str() );
-        }
-
-        for( const item &i : drop_item.contents ) {
-            g->m.add_item_or_charges( pt, i );
-        }
-        // TODO: Non-glass breaking
-        // TODO: Wine glass breaking vs. entire sheet of glass breaking
-        sounds::sound(pt, 16, _("glass breaking!"));
-        return;
-    }
-
-    // Copy the item
-    item dropped_item = drop_item;
-
-    monster *mon = dynamic_cast<monster *>( attack.hit_critter );
-
-    // We can only embed in monsters
-    bool embed = mon != nullptr && !mon->is_dead_state();
-    // And if we actually want to embed
-    embed = embed && effects.count( "NO_EMBED" ) == 0;
-    // Don't embed in small creatures
-    if( embed ) {
-        const m_size critter_size = mon->get_size();
-        const units::volume vol = dropped_item.volume();
-        embed = embed && ( critter_size > MS_TINY || vol < 250_ml );
-        embed = embed && ( critter_size > MS_SMALL || vol < 500_ml );
-        // And if we deal enough damage
-        // Item volume bumps up the required damage too
-        embed = embed &&
-                 ( attack.dealt_dam.type_damage( DT_CUT ) / 2 ) +
-                   attack.dealt_dam.type_damage( DT_STAB ) >
-                     attack.dealt_dam.type_damage( DT_BASH ) +
-                     vol * 3 / 250_ml + rng( 0, 5 );
-    }
-
-    if( embed ) {
-        mon->add_item( dropped_item );
-        if( g->u.sees( *mon ) ) {
-            add_msg( _("The %1$s embeds in %2$s!"),
-                     dropped_item.tname().c_str(),
-                     mon->disp_name().c_str() );
-        }
-    } else {
-        bool do_drop = true;
-        if( effects.count( "ACT_ON_RANGED_HIT" ) ) {
-            // Don't drop if it exploded
-            do_drop = !dropped_item.process( nullptr, attack.end_point, true );
-        }
-
-        if( do_drop ) {
-            g->m.add_item_or_charges( attack.end_point, dropped_item );
-        }
-
-        if( effects.count( "HEAVY_HIT" ) ) {
-            if( g->m.has_flag( "LIQUID", pt ) ) {
-                sounds::sound( pt, 10, _("splash!") );
-            } else {
-                sounds::sound( pt, 8, _("thud.") );
-            }
-            const trap &tr = g->m.tr_at( pt );
-            if( tr.triggered_by_item( dropped_item ) ) {
-                tr.trigger( pt, nullptr );
-            }
-        }
-    }
+    return dispersion;
 }
 
 double player::gun_value( const item &weap, long ammo ) const
@@ -2023,7 +1527,8 @@ double player::gun_value( const item &weap, long ammo ) const
 
     item tmp = weap;
     tmp.ammo_set( weap.ammo_default() );
-    int total_dispersion = get_weapon_dispersion( tmp, RANGE_SOFT_CAP ) + effective_dispersion( tmp.sight_dispersion() );
+    int total_dispersion = get_weapon_dispersion( tmp, RANGE_SOFT_CAP ).max() +
+      effective_dispersion( tmp.sight_dispersion() );
 
     if( def_ammo_i != nullptr && def_ammo_i->ammo != nullptr ) {
         const islot_ammo &def_ammo = *def_ammo_i->ammo;
